Move HGCGuardRing.h(cc) from SimG4CMS/Calo to Geometry/HGCalCommonData
