+++ Geometry/HGCalCommonData/BuildFile.xml
@@ -3,6 +3,7 @@
 <use name="DetectorDescription/Core"/>
 <use name="DetectorDescription/DDCMS"/>
 <use name="FWCore/ParameterSet"/>
+<use name="geant4core"/>
 <use name="dd4hep"/>
 <export>
   <lib name="1"/>
